requirements.txt
langchain
huggingface_hub
transformers
accelerate
bitsandbytes
langchain_huggingface
streamlit
python-dotenv